+++ codex/memory/__init__.py
@@ -1,4 +1,4 @@
-from .memory_store import save_memory, fetch_all, fetch_one, query
+from .memory_store import save_memory, fetch_all, fetch_one, query, search
 from .lineage import link_task_to_origin
 from . import doc_indexer
 
@@ -7,6 +7,7 @@ __all__ = [
     "fetch_all",
     "fetch_one",
     "query",
+    "search",
     "link_task_to_origin",
     "doc_indexer",
 ]